+++ index.html
@@ -14,11 +14,6 @@
     <body>
         <h2 class="header">Screenshots</h2>
         <div class="wrapper">
-            <div class="arrows-switch">
-                <p class="screen-readers">
-                    It's switch menu, that allows you to set visibility
-                </p>
-            </div>
             <h1 class="header header--main">Summer memories</h1>
         </div>
         <div class="container">
@@ -69,6 +64,6 @@
             </div>
             <div class="dotsNav"></div>
         </div>
-        <script defer src="src/main2.js"></script>
+        <script defer src="src/carousel.js"></script>
     </body>
 </html>
